remove segReaper thread from dataStor_ to idxMgr_
